# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shop Catalog API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserPublicRole =
  (typeof UserPublicRole)[keyof typeof UserPublicRole];

export const UserPublicRole = {
  admin: "admin",
  manager: "manager",
  user: "user",
} as const;

export interface UserPublic {
  id: number;
  username: string;
  role: UserPublicRole;
  isBlocked: boolean;
  hasActiveSession: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: UserPublic;
}

export interface ChangePasswordBody {
  currentPassword: string;
  newPassword: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  manager: "manager",
  user: "user",
} as const;

export interface CreateUserBody {
  username: string;
  password: string;
  role: CreateUserBodyRole;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  admin: "admin",
  manager: "manager",
  user: "user",
} as const;

export interface UpdateUserBody {
  username?: string;
  password?: string;
  role?: UpdateUserBodyRole;
}

export interface BlockUserBody {
  isBlocked: boolean;
}

export interface Catalog {
  id: number;
  name: string;
  parentId?: number | null;
  imageUrl?: string | null;
  sortOrder: number;
  createdAt: string;
  childCount: number;
  productCount: number;
}

export interface CatalogBreadcrumbItem {
  id: number;
  name: string;
}

export interface CreateCatalogBody {
  name: string;
  parentId?: number | null;
  imageUrl?: string | null;
  sortOrder?: number;
}

export interface UpdateCatalogBody {
  name?: string;
  imageUrl?: string | null;
  sortOrder?: number;
}

export interface MoveCatalogBody {
  newParentId?: number | null;
}

export interface ProductAttribute {
  key: string;
  value: string;
}

export interface Product {
  id: number;
  productId: string;
  name: string;
  price: number;
  catalogId: number;
  imageUrl?: string | null;
  images: string[];
  attributes: ProductAttribute[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateProductBody {
  name: string;
  price: number;
  catalogId: number;
  imageUrl?: string | null;
  images?: string[];
  attributes?: ProductAttribute[];
  productId?: string;
}

export interface UpdateProductBody {
  name?: string;
  price?: number;
  imageUrl?: string | null;
  images?: string[];
  attributes?: ProductAttribute[];
  productId?: string;
}

export interface MoveProductBody {
  newCatalogId: number;
}

export interface BulkDeleteBody {
  ids: number[];
}

export interface BulkMoveBody {
  ids: number[];
  newCatalogId: number;
}

export interface CartItemFull {
  id: number;
  productId: number;
  quantity: number;
  selectedColor?: string | null;
  product: Product;
  createdAt: string;
}

export interface AddToCartBody {
  productId: number;
  quantity: number;
  selectedColor?: string | null;
}

export interface UpdateCartItemBody {
  quantity?: number;
  selectedColor?: string | null;
}

export interface OrderItem {
  productId: number;
  productName: string;
  productCode: string;
  quantity: number;
  selectedColor?: string | null;
  price: number;
  imageUrl?: string | null;
}

export interface Order {
  id: number;
  userId?: number | null;
  guestName?: string | null;
  guestPhone?: string | null;
  items: OrderItem[];
  totalPrice?: number | null;
  status: string;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface PlaceOrderBody {
  guestName?: string;
  guestPhone?: string;
  notes?: string;
}

export interface UpdateOrderStatusBody {
  status: string;
  notes?: string;
}

export type ListCatalogsParams = {
  parentId?: number | null;
};

export type ListProductsParams = {
  catalogId: number;
};

export type ForceLogoutUser200 = {
  ok: boolean;
};

export type ListOrdersParams = {
  status?: string;
};