# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Shop Catalog API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Account blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "401":
          description: Unauthorized
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products in a catalog
      parameters:
        - name: catalogId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/move:
    post:
      operationId: moveProduct
      tags: [products]
      summary: Move product to another catalog
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveProductBody"
      responses:
        "200":
          description: Moved product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/bulk-delete:
    post:
      operationId: bulkDeleteProducts
      tags: [products]
      summary: Delete multiple products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteBody"
      responses:
        "204":
          description: Deleted
  /products/bulk-move:
    post:
      operationId: bulkMoveProducts
      tags: [products]
      summary: Move multiple products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkMoveBody"
      responses:
        "204":
          description: Moved
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current user cart
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItemFull"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add product to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "201":
          description: Added to cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItemFull"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear entire cart
      responses:
        "204":
          description: Cleared
  /cart/{id}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity or color
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItemFull"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /orders:
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderBody"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin/manager)
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin/manager)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserPublic"
      required: [user]
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, manager, user]
        isBlocked:
          type: boolean
        hasActiveSession:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, username, role, isBlocked, hasActiveSession, createdAt]
...
    ProductAttribute:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required: [key, value]
    Product:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: string
        name:
          type: string
        price:
          type: number
        catalogId:
          type: integer
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/ProductAttribute"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, productId, name, price, catalogId, images, attributes, createdAt, updatedAt]
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        catalogId:
          type: integer
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/ProductAttribute"
        productId:
          type: string
      required: [name, price, catalogId]
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            type: string
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/ProductAttribute"
        productId:
          type: string
    MoveProductBody:
      type: object
      properties:
        newCatalogId:
          type: integer
      required: [newCatalogId]
    BulkDeleteBody:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
      required: [ids]
    BulkMoveBody:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
        newCatalogId:
          type: integer
      required: [ids, newCatalogId]
    CartItemFull:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        selectedColor:
          type: string
          nullable: true
        product:
          $ref: "#/components/schemas/Product"
        createdAt:
          type: string
          format: date-time
      required: [id, productId, quantity, product, createdAt]
    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        selectedColor:
          type: string
          nullable: true
      required: [productId, quantity]
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
        selectedColor:
          type: string
          nullable: true
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        productCode:
          type: string
        quantity:
          type: integer
        selectedColor:
          type: string
          nullable: true
        price:
          type: number
        imageUrl:
          type: string
          nullable: true
      required: [productId, productName, productCode, quantity, price]
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          nullable: true
        guestName:
          type: string
          nullable: true
        guestPhone:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalPrice:
          type: number
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, items, status, createdAt, updatedAt]
    PlaceOrderBody:
      type: object
      properties:
        guestName:
          type: string
        guestPhone:
          type: string
        notes:
          type: string
      required: []
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
      required: [status]